+++ fastai/layers.py
@@ -2,7 +2,7 @@
 from .torch_core import *
 
 __all__ = ['AdaptiveConcatPool2d', 'BCEWithLogitsFlat', 'BCEFlat', 'MSELossFlat', 'CrossEntropyFlat', 'Debugger',
-           'Flatten', 'Lambda', 'PoolFlatten', 'ResizeBatch', 'bn_drop_lin', 'conv2d', 'conv2d_trans', 'conv_layer',
+           'Flatten', 'Lambda', 'PoolFlatten', 'View', 'ResizeBatch', 'bn_drop_lin', 'conv2d', 'conv2d_trans', 'conv_layer',
            'embedding', 'simple_cnn', 'NormType', 'relu', 'batchnorm_2d', 'trunc_normal_', 'PixelShuffle_ICNR', 'icnr',
            'NoopLoss', 'WassersteinLoss', 'SelfAttention', 'SequentialEx', 'MergeLayer', 'res_block', 'sigmoid_range',
            'SigmoidRange', 'PartialLayer', 'FlattenedLoss', 'BatchNorm1dFlat']
@@ -16,20 +16,32 @@ class Lambda(nn.Module):
 
     def forward(self, x): return self.func(x)
 
-def ResizeBatch(*size:int) -> Tensor:
-    "Layer that resizes x to `size`, good for connecting mismatched layers."
-    return Lambda(lambda x: x.view((-1,)+size))
+class View(nn.Module):
+    "Reshape `x` to `size`"
+    def __init__(self, *size:int):
+        super().__init__()
+        self.size = size
+
+    def forward(self, x): return x.view(self.size)
+
+class ResizeBatch(nn.Module):
+    "Reshape `x` to `size`, keeping batch dim the same size"
+    def __init__(self, *size:int):
+        super().__init__()
+        self.size = size
+
+    def forward(self, x):
+        size = (x.size(0),) + self.size
+        return x.view(size)
 
 class Flatten(nn.Module):
     "Flatten `x` to a single dimension, often used at the end of a model. `full` for rank-1 tensor"
     def __init__(self, full:bool=False):
         super().__init__()
         self.full = full
-    
+
     def forward(self, x):
         return x.view(-1) if self.full else x.view(x.size(0), -1)
-    #func = (lambda x: x.view(-1)) if full else (lambda x: x.view(x.size(0), -1))
-    #return Lambda(func)
 
 def PoolFlatten()->nn.Sequential:
     "Apply `nn.AdaptiveAvgPool2d` to `x` and then flatten the result."