+++ fastai/text/models/transformer.py
@@ -15,7 +15,7 @@ class PositionalEncoding(nn.Module):
         self.register_buffer('freq', 1 / (10000 ** (torch.arange(0., d, 2.)/d)))
     
     def forward(self, pos:Tensor, bs:int=None):
-        inp = torch.ger(pos, self.freq)
+        inp = torch.ger(pos.float(), self.freq)
         enc = torch.cat([inp.sin(), inp.cos()], dim=-1)
         return enc
 
@@ -205,6 +205,8 @@ class TransformerXL(nn.Module):
                 cat = torch.cat([self.hidden[i], hids[i]], dim=1)
                 self.hidden[i] = cat[:,-self.mem_len:].detach()
     
+    def select_hidden(self, idxs): self.hidden = [h[idxs] for h in self.hidden]
+    
     def forward(self, x):
         bs,x_len = x.size()
         inp = self.drop_emb(self.encoder(x)) #.mul_(self.d_model ** 0.5)
@@ -221,7 +223,7 @@ class TransformerXL(nn.Module):
             inp = layer(inp, r=pos_enc, u=self.u, v=self.v, mask=mask, mem=mem)
             hids.append(inp)
         core_out = inp[:,-x_len:]
-        self._update_mems(hids)
+        if self.mem_len > 0 : self._update_mems(hids)
         return (self.hidden if self.mem_len > 0 else [core_out]),[core_out]
 
 def init_transformer(m):
@@ -260,7 +262,7 @@ def tfmer_clas_split(model:nn.Module) -> List[nn.Module]:
 
 tfmerXL_lm_config = dict(ctx_len=150, n_layers=12, n_heads=10, d_model=410, d_head=41, d_inner=2100, resid_p=0.1, attn_p=0.1,
                          ff_p=0.1, embed_p=0.1, output_p=0.1, bias=False, scale=True, act=Activation.ReLU, double_drop=True,
-                         tie_weights=True, out_bias=False, init=init_transformer, mem_len=150)
+                         tie_weights=True, out_bias=True, init=init_transformer, mem_len=150)
 
 tfmerXL_clas_config = dict(ctx_len=150, n_layers=12, n_heads=10, d_model=410, d_head=41, d_inner=2100, resid_p=0.1, attn_p=0.1,
                          ff_p=0.1, embed_p=0.1, output_p=0.1, bias=False, scale=True, act=Activation.ReLU, double_drop=True,