+++ fastai/text/transform.py
@@ -32,8 +32,8 @@ class SpacyTokenizer(BaseTokenizer):
             self.tok.tokenizer.add_special_case(w, [{ORTH: w}])
 
 def spec_add_spaces(t:str) -> str:
-    "Add spaces around / and # in `t`."
-    return re.sub(r'([/#])', r' \1 ', t)
+    "Add spaces around / and # in `t`. \n"
+    return re.sub(r'([/#\n])', r' \1 ', t)
 
 def rm_useless_spaces(t:str) -> str:
     "Remove multiple spaces in `t`."
@@ -76,6 +76,7 @@ def deal_caps(x:Collection[str]) -> Collection[str]:
     "Replace all words in `x` by their lower version and add `TK_MAJ`."
     res = []
     for t in x:
+        if t == '': continue
         if t[0].isupper() and t[1:].islower(): res.append(TK_MAJ)
         res.append(t.lower())
     return res