+++ fastai/vision/image.py
@@ -93,12 +93,15 @@ class Image(ItemBase):
             return str_buffer.getvalue()
 
     def apply_tfms(self, tfms:TfmList, do_resolve:bool=True, xtra:Optional[Dict[Callable,dict]]=None,
-                   size:Optional[Union[int,TensorImageSize]]=None, resize_method:ResizeMethod=ResizeMethod.CROP,
+                   size:Optional[Union[int,TensorImageSize]]=None, resize_method:ResizeMethod=None,
                    mult:int=None, padding_mode:str='reflection', mode:str='bilinear', remove_out:bool=True)->TensorImage:
         "Apply all `tfms` to the `Image`, if `do_resolve` picks value for random args."
         if not (tfms or xtra or size): return self
         xtra = ifnone(xtra, {})
         tfms = sorted(listify(tfms), key=lambda o: o.tfm.order)
+        default_rsz = ResizeMethod.SQUISH if (size is not None and is_listy(size)) else ResizeMethod.CROP
+        resize_method = ifnone(resize_method, default_rsz)
+        if resize_method <= 2: tfms = self._maybe_add_crop_pad(tfms)
         if do_resolve: _resolve_tfms(tfms)
         x = self.clone()
         x.set_sample(padding_mode=padding_mode, mode=mode, remove_out=remove_out)